dcp-845 resolve circular import, separate crawler classes into packages
