+++ exporter/graph/entity/supplemented_entity.py
