+++ exporter/graph/info/process.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from typing import List
+
+from exporter.metadata.resource import MetadataResource
+
+
+@dataclass
+class ProcessInfo:
+    process: MetadataResource
+    inputs: List[MetadataResource]
+    outputs: List[MetadataResource]
+    protocols: List[MetadataResource]
