+++ exporter/graph/link/link_set.py
@@ -1,7 +1,7 @@
 from typing import Union, Dict, List
 
-from exporter.graph.process_link import ProcessLink
-from exporter.graph.supplementary_file_link import SupplementaryFileLink
+from .process import ProcessLink
+from .supplementary_file import SupplementaryFileLink
 
 Link = Union[ProcessLink, SupplementaryFileLink]
